pyproject.toml
[project]
name = "sdcourse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "isort>=6.0.1",
    "python-multipart>=0.0.20",
    "ruff>=0.12.10",
    "uvicorn>=0.35.0",
]
